+++ Sources/BaseFoundationiOS/Protocols/CoreProtocols/LocalizableProtocol.swift
@@ -5,7 +5,7 @@
 //  Created by pablo borquez on 20/07/23.
 //
 
-import Foundation
+import SwiftUI
 
 public protocol LocalizableProtocol: RawValueDataSource {
 }
@@ -13,6 +13,10 @@ public protocol LocalizableProtocol: RawValueDataSource {
 extension LocalizableProtocol {
     public var translate: String {
         return NSLocalizedString(self.rawValue, comment: "")
+    }
+    
+    public var text: Text {
+        return Text(self.translate)
     }
     
     public func translate(with arguments: Self...) -> String {